+++ docs/table-render/.test/simple2.jsx
@@ -1,10 +1,5 @@
-/**
- * transform: true
- * defaultShowCode: true
- */
-
 import React from 'react';
-import { Table, Search, TableProvider, useTable } from 'table-render';
+import { Table, Search, withTable, useTable } from 'table-render';
 import { Tag, Space, Menu, message, Tooltip, Button } from 'antd';
 import { PlusOutlined, InfoCircleOutlined } from '@ant-design/icons';
 import request from 'umi-request';
@@ -36,15 +31,7 @@ const schema = {
 };
 
 const Demo = () => {
-  return (
-    <TableProvider>
-      <TableBody />
-    </TableProvider>
-  );
-};
-
-const TableBody = () => {
-  const { refresh, tableState }: any = useTable();
+  const { refresh, tableState } = useTable();
 
   const searchApi = params => {
     console.log('params >>> ', params);
@@ -201,4 +188,4 @@ const TableBody = () => {
   );
 };
 
-export default Demo;
+export default withTable(Demo);