+++ packages/form-render/src/index.js
@@ -48,6 +48,7 @@ function App({
   allCollapsed = false,
   onValuesChange,
   column,
+  removeHiddenData = false,
   ...rest
 }) {
   try {
@@ -64,6 +65,7 @@ function App({
     setValueByPath,
     getSchemaByPath,
     setSchemaByPath,
+    setSchema,
     setValues,
     getValues,
     resetFields,
@@ -86,6 +88,7 @@ function App({
     isSubmitting,
     formData,
     flatten,
+    showValidate, // 旧版折中升级方案里，旧的api的软兼容
   } = valuesThatWillChange;
 
   useEffect(() => {
@@ -97,6 +100,7 @@ function App({
         validateMessages,
         beforeFinish,
         onMount,
+        removeHiddenData,
       });
     } else {
     }
@@ -134,6 +138,7 @@ function App({
       readOnly,
       disabled,
       allCollapsed,
+      showValidate,
     }),
     [
       displayType,
@@ -146,6 +151,7 @@ function App({
       readOnly,
       disabled,
       allCollapsed,
+      showValidate,
     ]
   );
 
@@ -158,6 +164,7 @@ function App({
       setEditing,
       touchKey,
       resetFields,
+      setSchema,
       setErrorFields,
       removeErrorField,
       removeTouched,