+++ packages/form-render/src/processData.js
@@ -6,7 +6,7 @@ import {
 import { unset, get, set } from 'lodash-es';
 import { isObject, clone } from './utils';
 // 提交前需要先处理formData的逻辑
-export const processData = (data, flatten) => {
+export const processData = (data, flatten, removeHiddenData) => {
   // 1. bind 的处理
   let _data = transformDataWithBind(data, flatten);
 
@@ -14,7 +14,9 @@ export const processData = (data, flatten) => {
   _data = removeEmptyItemFromList(_data);
 
   // 3. 去掉 hidden = true 的元素
-  // _data = removeHiddenFromResult(_data, flatten);
+  if (removeHiddenData) {
+    _data = removeHiddenFromResult(_data, flatten);
+  }
 
   // 4. 去掉所有的 undefined
   _data = cleanEmpty(_data);
@@ -114,10 +116,15 @@ export const transformDataWithBind2 = (data, flatten) => {
       const { key, bind } = item;
       const temp = [];
       bind.forEach(b => {
-        temp.push(get(newData, b));
+        const bindValue = get(newData, b);
+        if (bindValue !== undefined) {
+          temp.push(bindValue);
+        }
         unset(newData, b);
       });
-      set(newData, key, temp);
+      if (temp.length > 0) {
+        set(newData, key, temp);
+      }
     });
   };
   handleBindData2(_data);