+++ packages/form-render/src/widgets/antd/urlInput.js
@@ -2,20 +2,68 @@ import React from 'react';
 import { Input } from 'antd';
 import { isUrl } from '../../utils';
 
-const TestNode = ({ value }) => {
+const UrlNode = ({ value, addonText = '测试链接' }) => {
   const useUrl = isUrl(value);
+
   if (useUrl) {
     return (
       <a target="_blank" href={value}>
-        测试链接
+        {addonText}
       </a>
     );
   }
-  return <div>测试链接</div>;
+
+  return <div>{addonText}</div>;
 };
 
-export default function imageInput({ value, ...rest }) {
+export default function UrlInput({
+  value,
+  prefix,
+  suffix,
+  addonText,
+  onChange,
+  ...rest
+}) {
+  let _value = value || '';
+
+  if (prefix) {
+    _value = _value.replace(prefix, '');
+  }
+
+  if (suffix) {
+    _value = _value.replace(suffix, '');
+  }
+
+  const handleChange = e => {
+    let _value = e.target.value;
+    if (!_value) {
+      onChange(_value);
+      return;
+    }
+    if (prefix) {
+      _value = prefix + _value;
+    }
+    if (suffix) {
+      _value = _value + suffix;
+    }
+    onChange(_value);
+  };
+
   return (
-    <Input value={value} addonAfter={<TestNode value={value} />} {...rest} />
+    <Input
+      value={_value}
+      prefix={prefix}
+      suffix={suffix}
+      onChange={handleChange}
+      addonAfter={
+        <UrlNode
+          value={value}
+          prefix={prefix}
+          suffix={suffix}
+          addonText={addonText}
+        />
+      }
+      {...rest}
+    />
   );
 }